requirements.txt
# CORE DJANGO
Django==5.1.5
asgiref==3.8.1
sqlparse==0.5.1

# DJANGO EXTENSIONS
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.4.0
django-guardian==2.4.0
whitenoise==6.7.0
python-decouple==3.8

# DATABASE
PyMySQL==1.1.1

# EXPORTACIONES
django-xhtml2pdf==0.0.3
xhtml2pdf==0.2.13
reportlab==4.2.0
pypdf==4.3.1
html5lib==1.1
arabic-reshaper==3.0.0
python-bidi==0.4.2
xlsxwriter==3.2.0

# IMÁGENES
Pillow==10.4.0

# CELERY
celery==5.4.0
django-celery-beat==2.7.0
redis==5.0.7

# PRODUCCIÓN
gunicorn==22.0.0